terraform/modules/translation_worker/src/requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-speech==2.21.0
google-cloud-translate==3.12.0
google-cloud-pubsub==2.18.4
google-cloud-firestore==2.13.1
google-cloud-logging==3.8.0
